Add the feature plus title to the grammer file
I wasn't able to seperate the two files.
When I made the keywords a LEXER rule the parser couldn't match them anymore and when I tried to keep it a Parser rule it also gave errors.
Something about implicit literal. I combined the two again en voila.
It isn't what I hoped to do but it seems this is the good way of doing making the gramer.