Refactor the methods for running the containers
This is to prepare for running example tables.
The tables must run in a loop for rows that are present within the example table.
Therefor the method RunStepContainer must be cleaned up for readability.
A new exception MissingVariables is introduces for both the required input variables as the required output variables.
In the future we can split this in incomming an outgoing exceptoins.